pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sublingo"
version = "0.1.0"
description = "CLI subtitle translation tool powered by LLMs"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "pysubs2>=1.6",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-httpx>=0.30",
]

[project.scripts]
sublingo = "sublingo.main:main"

[tool.setuptools.packages.find]
include = ["sublingo*"]
